Update Bike_Rental_System.py
I edited the flow of the rent_or_return(), create_account(), and login() functions to reference the User objects and their attributes rather than the list of lists which just held the user details. I initialized the master list of users, user_list, as a global variable and created an infinite while loop to run in the main block unless broken by the user in order to simulate the bike rental app being open and running until the user manually closes it.