Update Bike_Rental_System.py
I added an update to the .user_info attribute of the user object inside the rent_bike function so it would update the objects in the list to allow .user_info to be called later and reflect changes to the objects when they rent. I also deleted some print lines used for checking output while coding.