Update Bike_Rental_System.py
This includes the rental_check() function to determine if a user is currently renting a bike. It also includes the return_bike() function to replace the bike in the available_bikes list and determine the cost to the user. There is also the create_excel() function which creates an excel doc of the current users once the program is finished. There are also some error checking changes to make sure the program does not get stuck or variables aren't called before they are initialized.